Refractor biome/structure/gen code and add new/better organization. Remove useless world gen API
