+++ src/api/java/de/teamlapen/vampirism/api/VampirismAPI.java
@@ -13,7 +13,6 @@ import de.teamlapen.vampirism.api.entity.player.vampire.IVampirePlayer;
 import de.teamlapen.vampirism.api.entity.player.vampire.IVampireVisionRegistry;
 import de.teamlapen.vampirism.api.items.IExtendedBrewingRecipeRegistry;
 import de.teamlapen.vampirism.api.world.IVampirismWorld;
-import de.teamlapen.vampirism.api.world.IWorldGenConfiguration;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
@@ -43,7 +42,6 @@ public class VampirismAPI {
     private static ISkillManager skillManager;
     private static IActionManager actionManager;
     private static IEntityActionManager entityActionManager;
-    private static IWorldGenConfiguration worldGenRegistry;
     private static IExtendedBrewingRecipeRegistry extendedBrewingRecipeRegistry;
 
     public static ISkillManager skillManager() {
@@ -83,12 +81,6 @@ public class VampirismAPI {
         return entityRegistry;
     }
 
-    /**
-     * @return The world gen registry
-     */
-    public static IWorldGenConfiguration worldGenRegistry() {
-        return worldGenRegistry;
-    }
 
     public static IExtendedBrewingRecipeRegistry extendedBrewingRecipeRegistry() {
         return extendedBrewingRecipeRegistry;
@@ -101,7 +93,7 @@ public class VampirismAPI {
      * @throws IllegalStateException if the API was already setup
      */
     public static void setUpRegistries(IFactionRegistry factionRegistryIn, ISundamageRegistry sundamageRegistryIn, IVampirismEntityRegistry entityRegistryIn, IActionManager actionManagerIn, ISkillManager skillManagerIn,
-                                       IVampireVisionRegistry vampireVisionRegistryIn, IEntityActionManager entityActionManagerIn, IWorldGenConfiguration worldGenRegistryIn, IExtendedBrewingRecipeRegistry extendedBrewingRecipeRegistryIn) {
+                                       IVampireVisionRegistry vampireVisionRegistryIn, IEntityActionManager entityActionManagerIn, IExtendedBrewingRecipeRegistry extendedBrewingRecipeRegistryIn) {
         if (INIT) throw new IllegalStateException("Vampirism API can only be setup once");
         factionRegistry = factionRegistryIn;
         sundamageRegistry = sundamageRegistryIn;
@@ -110,7 +102,6 @@ public class VampirismAPI {
         skillManager = skillManagerIn;
         vampireVisionRegistry = vampireVisionRegistryIn;
         entityActionManager = entityActionManagerIn;
-        worldGenRegistry = worldGenRegistryIn;
         extendedBrewingRecipeRegistry = extendedBrewingRecipeRegistryIn;
         INIT = true;
     }
