+++ src/main/java/de/teamlapen/vampirism/mixin/SurfaceRuleDataMixin.java
@@ -1,8 +1,8 @@
 package de.teamlapen.vampirism.mixin;
 
 import com.google.common.collect.ImmutableList;
-import de.teamlapen.vampirism.core.ModBiomes;
 import de.teamlapen.vampirism.modcompat.terrablender.TerraBlenderCompat;
+import de.teamlapen.vampirism.world.biome.OverworldModifications;
 import net.minecraft.data.worldgen.SurfaceRuleData;
 import net.minecraft.world.level.levelgen.SurfaceRules;
 import org.spongepowered.asm.mixin.Mixin;
@@ -15,7 +15,7 @@ public abstract class SurfaceRuleDataMixin {
     @ModifyVariable(method = "overworldLike(ZZZ)Lnet/minecraft/world/level/levelgen/SurfaceRules$RuleSource;", at = @At("STORE"), ordinal = 0)
     private static ImmutableList.Builder<SurfaceRules.RuleSource> addVampirismOverworldSurfaceRules(ImmutableList.Builder<SurfaceRules.RuleSource> builder){
         if(!TerraBlenderCompat.areBiomesAddedViaTerraBlender()){ //When TerraBlender is installed, it adds the surface rules appropriately. This is likely called a few times before terrablender is activated, but that should not be an issue
-            builder.add(ModBiomes.buildOverworldSurfaceRules());
+            builder.add(OverworldModifications.buildOverworldSurfaceRules());
         }
         return builder;
     }
