+++ src/main/java/de/teamlapen/vampirism/world/gen/util/BiomeTopBlockProcessor.java
@@ -1,7 +1,7 @@
 package de.teamlapen.vampirism.world.gen.util;
 
 import com.mojang.serialization.Codec;
-import de.teamlapen.vampirism.core.ModFeatures;
+import de.teamlapen.vampirism.world.gen.VampirismFeatures;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
@@ -48,7 +48,7 @@ public class BiomeTopBlockProcessor extends StructureProcessor {
     @Nonnull
     @Override
     protected StructureProcessorType<?> getType() {
-        return ModFeatures.biome_based;
+        return VampirismFeatures.biome_based;
     }
 
 }
